fix: The holy molly checker

+++ python_variable_annotations/6-sum_mixed_list.py
@@ -5,7 +5,7 @@ a list of integers and floating-point numbers."""
 from typing import List, Union
 
 
-def sum_mixed_list(nums: List[Union[float, int]]) -> float:
+def sum_mixed_list(nums: List[Union[int, float]]) -> float:
     """
     Sums a list of mixed numbers (integers and floats).
